# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start API",
        "description": "AgroBozor uchun backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.daladan.uz/api/v1/",
            "description": "Production server"
        },
        {
            "url": "http://daladan-api.loc/api/v1",
            "description": "Local dev server (OSPanel)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "php artisan serve"
        }
    ],
    "paths": {
...
        "/resources/cities": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Faol tumanlar ro'yxati (ixtiyoriy region_id bo'yicha)",
                "description": "cities() — GET /resources/cities",
                "operationId": "0d41ec81256a3feecd7b0d99e59fabe4",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Faqat shu viloyatga tegishli tumanlar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tumanlar ro'yxati",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResourceCity"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/resources/subcategories": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Faol subkategoriyalar ro'yxati (ixtiyoriy category_id bo'yicha)",
                "description": "subcategories() — GET /resources/subcategories",
                "operationId": "d29b942b23445d27362e2eb4911c49dd",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Faqat shu kategoriyaga tegishli subkategoriyalar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subkategoriyalar ro'yxati",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResourceSubcategory"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResourceCity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "region_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name_uz": {
                        "type": "string",
                        "example": "Qorovulbozor"
                    },
                    "slug": {
                        "type": "string",
                        "example": "qorovulbozor"
                    }
                },
                "type": "object"
            },
...
            "ResourceSubcategory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 11
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "name": {
                        "type": "string",
                        "example": "Echkilar"
                    },
                    "slug": {
                        "type": "string",
                        "example": "echkilar"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}